requirements.txt
mysql-connector-python==8.1.0
customtkinter==5.2.1
Pillow==10.1.0
packaging==23.2
